cmake: revamp lapack and blas libraries usage
Since lapack is likely to depend on blas, link with -llapack -lblas
to correctly support lapack and/or blas static libraries.

Also, add the option to append extra libraries via the
GMX_LAPACK_EXTRA and GMX_BLAS_EXTRA macros. For example, the latter
can be set to -lgfortran if blas is provided as a static library.