Turn InteractionList into a class
Made all data members private.
Store functionType. This enables removing separate passing of the
function type to functions in a later change.
Add iterators.
Also changes ilist in moltype and some other places from an array
of InteractionList to InteractionLists.
Removed InteractionListHandle now it is no longer needed and
replaced it by InteractionListsSelection to loop over selections
of InteractionLists.

Fixed a bug in the MiMiC preprocessing, which would always add
the parameters of atoms of the first bond to a QM list instead
of different bonds (this looks like it was never tested).

TODO: Add unit tests