Add data type size asserts in the nbnxm OpenCL module
This commits adds static_asserts to make sure that data transfers
to/from the GPU with OpenCL happen between compatible host and
device data types.
Assertions are also added for non-struct parameters passed by value in
the nbnxm module (size structs can not be checked in a simple way).

Implements part of #2936