Split up logging streams
The MDLogger object and builder have been changed to split up the
previous logging stream organization to instead be based on a collection
of logging streams and verbosity levels for each stream.

For now, 4 different streams (Error, Warning, Info, Debug) are defined,
which each of them having either Verbose or NoVerbose logging.

When building the logging object, different text streams can be
associated with the different output targets and verbosity levels.
Messages at verbose level are not printed if the registered stream is
only supposed to write at lower verbosity level instead, but higher
verbosity messages are always printed to lower verbosity streams as
well.

The GMX_LOG macro has been modified to now log to the low verbosity
output of the given message stream, as to not change too many lines of
code here. A new macro that allows logging to the verbose level has been
added, as well as a generic version that allows logging to any
verbosity.

Refs #3013

Change-Id: I78313375e8694f909a85f41167f0f1a739e56d03