WIP: Support icc in gitlab CI
Generalizes the TSAN build stage to a compiler build stage, and uses
it for containing the installation process for the intel
compiler. When installing intel, a gcc is needed also, and this is
configurable also.

TODO Decide how we're going to bind mount the license file
securely into the configure and build containers.

TODO Document the approach once it works

Refs #3459