Modernize GitLab CI rules.
Convert to the newer *rules* syntax for consistency, flexibility, and
future proofing, since there are warnings about future depcrecation of
*only* and *except*.

Remove some *job:when* parameters that escaped the previous
consolidation of job selection logic. This may prevent jobs that would
otherwise have run in some edge cases, but such logic was undocumented
and will now be easier to document and troubleshoot.