Make source linters compare against HEAD in scripts
Changes the default behaviour back to compare against the current HEAD
in the source linter scripts for copyright and clang-format. CI will
now compare against origin/master by default, but this should likely be
changed to be dependent on the type of pipeline being run, so that
nightly and release pipelines will compare against HEAD^ instead.

Refs #3463

Change-Id: I3bf6a944c3e3ec44841243cf3d1aef5e897d37f6