Remove extraneous override of default job dependencies.
The GitLab documentation is not clear on the expected behavior, but it
seems like setting an empty *dependencies* array may be preventing
preceding stage failure from causing early failure of the pipeline,
rendering the pre-build stage ineffective. There are not substantial
artifacts to prevent loading, so removing the explicit
`dependencies: []` seems harmless.