Allow updates of multichoice CMake variables
This refactors the routines for multichoice and trivalue
CMake variables so the actual cache setting is done in
a separate lowlevel routine that can also be called later
to update the value, in particular changing AUTO to a
specific (detected) choice. This will help us avoid
having separate variables for the original setting and
the currently active setting.

The previous string case handling was fragile. The values
were stored as arbitrary case, but upon returning the
parent scope was set to an upper case version. However, the
cache would not be updated, so anything beyond the parent
scope would have a different-case version. To simplify, we
now store the values in the actual cases provided.