Template size dependent GPU datastructures
The GPU datastructures depending on the c_nbnxnGpuClusterSize static
variable have been templated to select the correct size depending on the
type of hardware it is run, with default 8 for CUDA and AMD OpenCL, and
4 for Intel OpenCL. This removes the need for recompiling a GROMACS
OpenCL build to run on Intel OpenCL hardware if it was build before for
AMD. It also prepares for later refactoring by removing hardcoded
variable sizes in both the CUDA and OpenCL kernel datastructures. It is
in theory also possible to use the smaller cluster size for CUDA, but
that is not used right now.

Change-Id: Ided4b83e14fb469dfb18bc47560a271d912a737b