Fix pbcutil module
Removed dependency of installed header on pbc.h, since it was easy.
The inclusion was only necessary for a default initialization of
PbcType, which is not strictly needed either. If we're going to offer
people a legacy API then we have to make compromises, and I think this
is a better one than installing an additional header.