Refactor virial calculation and move to correct translation unit
This moves the previously static virial calculation function from
sim_util to the correct translation unit and also renames the
functions there. One overload that provided no utility is also
removed. The call signature of the low-level virial calculation
is modernized and the, now deprecated, legacy implementation
is moved to be internal to the virial translation unit to
prevent further use.