Address clang-tidy issues
Fixed issues include a misplaced typdef const, using RAII for
file IO types.

Added two nolints. The first is for a clang-tidy-9 bug where
constexpr brackets get mangled. Should be fixed in clang-tidy-11.

Second NOLINT is for a static member accessed through instance. This
is unrelated to the MR and will likely be fixed alongside work
on the parallel trajectory framework