Factor out fit index selection.
Fit indices are now computed in a function that does not modify
the user input.

Also fixed a tiny bug - now the gmx msd output reports taus that
were ACTUALLY fit to. When beginFit does not match the step of dt,
it previously reported the desired fit, not the rounded fit